modification

+++ src/main/java/com/apicollabdev/odk/collabdev/entity/Fonctionnalite.java
@@ -14,7 +14,7 @@ public class Fonctionnalite {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_fonctionnalite")
-    private int idFonctionnalite;
+    private Long idFonctionnalite;
 
     private int pointFonctionnalite;
     @Enumerated(EnumType.STRING)
@@ -44,11 +44,12 @@ public class Fonctionnalite {
     @JsonManagedReference
     private Contribution contribution;
 
-    public int getIdFonctionnalite() {
+
+    public Long getIdFonctionnalite() {
         return idFonctionnalite;
     }
 
-    public void setIdFonctionnalite(int idFonctionnalite) {
+    public void setIdFonctionnalite(Long idFonctionnalite) {
         this.idFonctionnalite = idFonctionnalite;
     }
